fix(logger): add filter to allow root handlers
Add a common filter to the application log handlers to allow these to be
moved to the root of the logging hierarchy without being swamped by
unwanted pds4_tools messages (issue open in that repo to try to get it
to honour its quietness in terms of log propagation in the future).

The motivation for moving the loggers to the root of the hierarchy is to
allow us to incorporate messages from libraries like Passthrough.