build(cli): increase modularity with opt. deps.
Set `typer` as an optional dependency (extras = "cli"), to prevent
consuming applications from having to pull in this (and by extension
`click`) if they don't care about the `cli` subpackage.

Warn users who try to import the `cli` subpackage without having
installed the extras.

Not entirely happy about the import guard implementation, but it be
good enough for now (tm).