feat(products): expand meta maps
Standardise timestamp monikers to follow `[start|stop]_{format}`,
but leave `start` and `stop` as aliases for the UTC ones.

Was intending to include a `vid_modified` moniker which would resolve
the `modification_date` of the `Modification_Detail` associated with the
current VID, but it looks like the ElementTree labels that pds4_tools
yields lack the XPath predicate features necessary to target the VID
dynamically. `[child="literal"]` is supported, but
`[child=/path/to/elem]` or similar is not.