Lots of work on cli/status.py
After pondering overnight, I understand what it's trying to do and
why it doesn't just use an Enum. Python Enum's can't be subclassed,
and Ariel wanted something that behaves like an enum but with
subclassing, so he could define some base exit code but allow
proctools users to add to the list.

I've added lots of comments and (hopefully) improved the checks
that __init_subclass__ does.